Modularize babel config
